+++ NEWS.md
@@ -1,6 +1,6 @@
 ## simplexR 0.0.0.9002
 
-Added big-M method to handle arbitrary linear programms.
+Added big-M method to handle arbitrary linear programs.
 
 ## simplexR 0.0.0.9001