+++ NEWS.md
@@ -1,3 +1,7 @@
+## simplexR 0.0.0.9003
+
+Added dual simplex algorithm and revised simplex algorithm.
+
 ## simplexR 0.0.0.9002
 
 Added big-M method to handle arbitrary linear programs.